Move oauth server to new package
