+++ oauth/oauth_server.go
@@ -1,4 +1,4 @@
-package config
+package oauth
 
 import (
 	"context"
